requirements.txt
certifi==2020.4.5.1
cffi==1.14.0
cryptography==2.9.2
editdistance==0.5.3
epitran==1.8
marisa-trie==0.7.5
munkres==1.1.2
nltk==3.4.5
numpy==1.18.1
pandas==0.25.3
panphon==0.17
pyaes==1.6.1
pycparser==2.20
pydub==0.23.1
pyenchant==2.0.0
python-dateutil==2.8.1
python-speech-features==0.6
pytz==2019.3
PyYAML==5.3
regex==2020.2.20
scipy==1.4.1
six==1.14.0
textblob==0.15.3
textblob-fr==0.2.0
unicodecsv==0.14.1
